requirements.txt
aiofiles==0.5.0
aiosqlite==0.15.0
appdirs==1.4.4
asgiref==3.2.10
asyncpg==0.21.0
attrs==20.2.0
bcrypt==3.2.0
black==19.10b0
cffi==1.14.2
click==7.1.2
colorama==0.4.3
databases==0.3.2
dnspython==2.0.0
docstring-parser==0.7.1
email-validator==1.1.1
fastapi==0.61.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
httptools==0.1.1
Hypercorn==0.10.2
hyperframe==5.2.0
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==3.3.1
pathspec==0.8.0
piccolo==0.12.6
piccolo-admin==0.9.2
piccolo-api==0.10.1
pkg-resources==0.0.0
priority==1.3.0
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
python-multipart==0.0.5
regex==2020.7.14
secure==0.2.1
six==1.15.0
starlette==0.13.6
starlette-json==19.11.15
targ==0.1.6
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.2.0
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
wsproto==0.15.0
WTForms==2.3.3
